refactor: rename `Marketplaces` component to `MarketplaceList` (#671)
